+++ cocsim/buildings/bomb_tower.py
@@ -68,7 +68,7 @@ class BombTower(SplashProjectileActiveBuilding):
     def projectile_speed(self) -> float:
         return 8.0
 
-    def __init__(self, game: "game.Game", x: float, y: float, level: int):
+    def __init__(self, game: "game.Game", x: int, y: int, level: int):
         super().__init__(
             game,
             x,
