+++ cocsim/buildings/cannon.py
@@ -68,7 +68,7 @@ class Cannon(ProjectileActiveBuilding):
     def projectile_speed(self) -> float:
         return 12.0
 
-    def __init__(self, game: "game.Game", x: float, y: float, level: int):
+    def __init__(self, game: "game.Game", x: int, y: int, level: int):
         super().__init__(
             game,
             x,
