+++ cocsim/buildings/wizard_tower.py
@@ -71,7 +71,7 @@ class WizardTower(SplashProjectileActiveBuilding):
 
         return 5.0
 
-    def __init__(self, game: "game.Game", x: float, y: float, level: int):
+    def __init__(self, game: "game.Game", x: int, y: int, level: int):
         super().__init__(
             game,
             x,
